+++ src/index.ts
@@ -1,26 +1,21 @@
-// Source mapping
-if (!process[Symbol.for('ts-node.register.instance')]) {
-	// eslint-disable-next-line @typescript-eslint/no-var-requires
-	require('source-map-support').install({
-		environment: 'node'
-	});
-}
+/**
+ * This file is the main entrypoint for your application. It just runs the
+ * 	register function (read the readme), and then runs the main function.
+ *
+ * You should typically avoid changing this file if you don't have to, to
+ * 	avoid merge conflicts with the template.
+ */
+import { bootstrap } from 'ts-async-bootstrap';
+import { errorHandler } from './error-handler';
+import { main } from './main';
+import { register } from './register';
 
-// Stack traces
-Error.stackTraceLimit = Infinity;
-const nativePrepareStackTrace = Error.prepareStackTrace;
-Error.prepareStackTrace = (err, traces) => {
-	return nativePrepareStackTrace(err, traces)
-		.split('\n')
-		.filter(line => {
-			return (
-				line &&
-				!line.includes('node_modules') &&
-				!line.includes('internal/process/')
-			);
-		})
-		.join('\n');
-};
-
-// Enter application
-import './main';
+/**
+ * Bootstrap the application
+ */
+bootstrap({
+	register: register,
+	run: main,
+	errorHandler: errorHandler,
+	shouldExit: true
+});