+++ src/main.ts
@@ -1,11 +1,9 @@
-import { logger } from './logger';
+import { log } from './log';
 import { env } from './environment';
 
-async function main() {
-	logger.info('Hello ' + env.APP_TITLE);
+/**
+ * Start your application in the main() function
+ */
+export async function main(): Promise<void> {
+	log.info('Hello ' + env.APP_TITLE);
 }
-main()
-	.catch(error => {
-		logger.error(error);
-		process.exit();
-	});