+++ test/spec/log.spec.ts
@@ -1,0 +1,7 @@
+import { log } from '../../src/log';
+
+describe('Log', () => {
+	it('exports an instance', () => {
+		expect(log).toBeDefined();
+	});
+});